# swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Spiritory Swagger documentation",
    "version": "1.0.0",
    "title": "Spiritory",
    "termsOfService": "http://swagger.io/terms/",

    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "host": "localhost:3033",
  "basePath": "/api",
  "tags": [
    {
      "name": "login",
      "description": "Everything about your login",
      "externalDocs": {
        "description": "Find out more",
        "url": "http://swagger.io"
      }
    },
    {
      "name": "User",
      "description": "Post and Get User",
      "externalDocs": {
        "description": "Find out more about our store",
        "url": "http://swagger.io"
      }
    },
    {
      "name": "Purchase",
      "description": "Get user's purchases",
      "externalDocs": {
        "description": "Find out more about our store",
        "url": "http://swagger.io"
      }
    }
  ],
  "schemes": ["http"],
  "paths": {
    "/login": {
      "post": {
        "tags": ["login"],
        "summary": "User login",
        "description": "User login",
        "operationId": "login",
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "login",
            "required": true,
            "schema": {
              "$ref": "#/definitions/login"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation"
          },
          "400": {
            "description": "Password incorrect"
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
    "/users/add": {
      "post": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "tags": ["User"],
        "summary": "Create User",
        "description": "Add user",
        "operationId": "addUser",
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Create user",
            "required": true,
            "schema": {
              "$ref": "#/definitions/addUser"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation"
          },
          "400": {
            "description": "Email exisit"
          },
          "404": {
            "description": "Empty content"
          }
        }
      }
    },
    "/users/find/{id}": {
      "get": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "tags": ["User"],
        "summary": "Get user by id ",
        "description": "Find user by id ",
        "operationId": "getUserById",
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "ID of body",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation"
          },
          "404": {
            "description": "Not found "
          }
        }
      }
    },
    "/purchases/getPurchaseByUser/{userId}": {
      "get": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "tags": ["Purchase"],
        "summary": "Get user's purchases ",
        "description": "Find all user's purchases",
        "operationId": "getPurchaseByUser",
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "path",
            "name": "userId",
            "description": "ID of body",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation"
          },
          "404": {
            "description": "Not found "
          }
        }
      }
    }
  },
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  },
  "definitions": {
    "login": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      }
    },
    "addUser": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "lastName": {
          "type": "string"
        },
        "birthday": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      }
    }
  },
  "externalDocs": {
    "description": "Find out more about Swagger",
    "url": "http://swagger.io"
  }
}